+++ lib/inventory/admin.dart
@@ -9,8 +9,8 @@ import 'package:beaute_app/inventory/stock/categories/views/categories.dart';
 import 'package:beaute_app/inventory/sellpoint/cart/views/cart.dart';
 import 'package:beaute_app/inventory/scanBarCode.dart';
 import 'package:beaute_app/inventory/stock/products/views/products.dart';
+import 'package:beaute_app/inventory/testPrinter/printConnections.dart';
 import 'package:beaute_app/inventory/testPrinter/printService.dart';
-import 'package:beaute_app/inventory/testPrinter/test2.dart';
 import 'package:beaute_app/navBar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
