+++ lib/inventory/listenerPrintService.dart
@@ -1,6 +1,6 @@
 
 
-import 'package:beaute_app/inventory/testPrinter/printService.dart';
+import 'package:beaute_app/inventory/testPrinter/printConnections.dart';
 
 class ListenerPrintService {
   int _printServiceActivated = 0;
