+++ lib/inventory/sellpoint/cart/views/cart.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'package:beaute_app/inventory/sellpoint/cart/services/cartService.dart';
 import 'package:beaute_app/inventory/sellpoint/cart/styles/cartStyles.dart';
-import 'package:beaute_app/inventory/testPrinter/test2.dart';
 import 'package:beaute_app/inventory/testPrinter/printService.dart';
+import 'package:beaute_app/inventory/testPrinter/printConnections.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
