+++ lib/inventory/testPrinter/printConnections.dart
@@ -1,0 +1,138 @@
+
+import 'dart:async';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+import 'dart:typed_data';
+import '../../agenda/utils/showToast.dart';
+import '../../agenda/utils/toastWidget.dart';
+import '../listenerPrintService.dart';
+
+
+class PrintService extends ChangeNotifier {
+  @override
+    void dispose() {
+      _connectionSubscription?.cancel();
+      super.dispose();
+    }
+
+  FlutterBlue flutterBlue = FlutterBlue.instance;
+  List<BluetoothDevice> devicesList = [];
+  BluetoothDevice? selectedDevice;
+  BluetoothCharacteristic? characteristic;
+  String targetDevice = '0ED2DB67-8733-2C1D-0ACB-557F656FFCF3';
+  StreamSubscription<BluetoothDeviceState>? _connectionSubscription;
+  bool isConnected = false;
+  ListenerPrintService listenerPrintService = ListenerPrintService();
+  String nameTargetDevice = 'MP210';
+  Future<void> ensureCharacteristicAvailable() async {
+    if (characteristic == null && selectedDevice != null) {
+      await discoverServices(selectedDevice!);
+    }
+    int retryCount = 0;
+    while (characteristic == null && retryCount < 10) {
+      await Future.delayed(Duration(milliseconds: 500));
+      retryCount++;
+    }
+    if (characteristic == null) {
+      throw Exception("Error: Característica de impresión no disponible");
+    }
+  }
+  Future<void> discoverServices(BluetoothDevice device) async {
+    List<BluetoothService> services = await device.discoverServices();
+    for (BluetoothService service in services) {
+      for (BluetoothCharacteristic charac in service.characteristics) {
+        if (charac.properties.write) {
+          characteristic = charac;
+          notifyListeners();
+          return;
+        }
+      }
+    }
+  }
+
+    void initDeviceStatus() {
+      isConnected = selectedDevice != null;
+      selectedDevice !=null ? listenerPrintService.setChange(3) : null;
+    }
+
+
+  void scanForDevices(context) async {
+    try {
+      List<BluetoothDevice> connectedDevices = await flutterBlue.connectedDevices;
+      for (BluetoothDevice device in connectedDevices) {
+        if (device.id.toString() == targetDevice) {
+          selectedDevice = device;
+          disconnect(context);
+          /*discoverServices(selectedDevice!);
+          listenToDeviceState(context);
+          notifyListeners();*/
+          return;
+        }
+      }
+      flutterBlue.startScan(timeout: const Duration(seconds: 5));
+      flutterBlue.scanResults.listen((results) async {
+        for (ScanResult r in results) {
+          if (r.device.name == nameTargetDevice) {
+            print("name ${r.device.name}");
+            flutterBlue.stopScan();
+            selectedDevice = r.device;
+            print("Dispositivo encontrado: ${selectedDevice?.name}");
+            try {
+              await Future.delayed(const Duration(seconds: 2));
+              await selectedDevice?.connect();
+              isConnected = true;
+              listenerPrintService.setChange(1);
+              print("Dispositivo conectado: ${selectedDevice?.name}");
+              discoverServices(selectedDevice!);
+              showOverlay(context, const CustomToast(message: "Dispositivo conectado correctamente"));
+              listenToDeviceState(context);
+              notifyListeners();
+            } catch (e) {
+              print("Error al conectar con el dispositivo: $e");
+              selectedDevice = null;
+              showOverlay(context, const CustomToast(message: "Espere mientras se reconecta automáticamente"));
+              await Future.delayed(const Duration(seconds: 8));
+              scanForDevices(context);
+              notifyListeners();
+            }
+            break;
+          }}});
+    } catch (e) {
+      print("Error durante el escaneo: $e");
+    }}
+
+  void listenToDeviceState(context) {
+    _connectionSubscription?.cancel();
+    if (selectedDevice != null) {
+      _connectionSubscription = selectedDevice!.state.listen((state) {
+        if (state == BluetoothDeviceState.disconnected) {
+          selectedDevice = null;
+          showOverlay(context, const CustomToast(message: 'Impresora desconectada'));
+          notifyListeners();
+        }});}}
+  void disconnect(context) async {
+    if (selectedDevice != null) {
+      try {
+        _connectionSubscription?.cancel();
+        await selectedDevice?.disconnect();
+        selectedDevice = null;
+        isConnected = false;
+        listenerPrintService.setChange(0);
+        notifyListeners();
+        print("Dispositivo desconectado correctamente");
+        showOverlay(context, const CustomToast(message: 'Dispositivo desconectado correctamente'));
+      } catch (e) {
+        print("Error al desconectar el dispositivo: $e");
+      }
+    }
+  }
+  Future<Uint8List> loadImageFromFile(String path) async {
+    try {
+      return await rootBundle.load(path).then((byteData) => byteData.buffer.asUint8List());
+    } catch (e) {
+      throw Exception("Error al cargar la imagen: $e");
+    }
+  }
+
+}
